requirements.txt
fastapi
uvicorn[standard]
ultralytics
pillow
python-dotenv
pyngrok
sahi
opencv-python-headless
numpy
python-multipart
Flask
requests
Pillow
numpy
